Don't let the tornado destroy so much blocks
The tornado now only destroys a few blocks and simply moves through
walls. The problem is, that in walls the tornado still spawns falling
blocks IN the walls, causing thousands of entities like drops to be
spawned which really overflows the client (leading to laggs, timeouts or
even crashes).
TODO: Find a solution for this.

* Reduced amount of blocks in the tornado
* Don't use the changed-block list anymore (not very useful)